+++ tools/dataspace_s1_download.py
@@ -13,28 +13,19 @@ import argparse
 import json
 import os
 from concurrent.futures import ThreadPoolExecutor, as_completed
-from datetime import datetime, timedelta
+from datetime import datetime
 from getpass import getuser
 from pathlib import Path
-from threading import Lock
-from typing import Tuple
 
 import backoff
 import requests
 
-from commons.logger import logger, LogLevels
-
-DEFAULT_QUERY_ENDPOINT = 'https://catalogue.dataspace.copernicus.eu/odata/v1/Products'
-"""Default URL endpoint for the Copernicus Data Space Ecosystem (CDSE) query REST service"""
-
-DEFAULT_AUTH_ENDPOINT = 'https://identity.dataspace.copernicus.eu/auth/realms/CDSE/protocol/openid-connect/token'
-"""Default URL endpoint for performing user authentication with CDSE"""
-
-DEFAULT_DOWNLOAD_ENDPOINT = 'https://download.dataspace.copernicus.eu/odata/v1/Products'
-"""Default URL endpoint for CDSE download REST service"""
-
-DEFAULT_SESSION_ENDPOINT = 'https://identity.dataspace.copernicus.eu/auth/realms/CDSE/account/sessions'
-"""Default URL endpoint to manage authentication sessions with CDSE"""
+from commons.logger import logger
+from util.backoff_util import fatal_code, backoff_logger
+from util.dataspace_util import (DEFAULT_QUERY_ENDPOINT,
+                                 DEFAULT_DOWNLOAD_ENDPOINT,
+                                 NoQueryResultsException,
+                                 DataspaceSession)
 
 ISO_TIME = '%Y-%m-%dT%H:%M:%SZ'
 """Temporal format required by ODATA API: yyyy-mm-ddTHH:MM:SSZ"""
@@ -43,120 +34,9 @@ QUERY_PAGE_SIZE = 1000
 """Max number of files returned per query"""
 
 
-def fatal_code(err: requests.exceptions.RequestException) -> bool:
-    """Only retry for common transient errors"""
-    return err.response.status_code not in [418, 429, 500, 502, 503, 504]
-
-
 def to_datetime(value) -> datetime:
     """Helper function to covert command-line arg to datetime object"""
     return datetime.strptime(value, ISO_TIME)
-
-
-class NoQueryResultsException(Exception):
-    """Custom exception to identify empty results from a query"""
-    pass
-
-
-class DataspaceSession:
-    """
-    Context manager class to wrap credentialed operations in.
-    Creates a session and gets its token on entry, and deletes the session on exit.
-    """
-
-    def __init__(self, username, password):
-        self.__lock = Lock()
-        self.__token, self.__session, self.__refresh_token, self.__expires = self._get_token(username, password)
-
-    @property
-    def token(self) -> str:
-        """Get the current access token for the Dataspace session. If it is expired, it will attempt to refresh"""
-
-        with self.__lock:
-            if datetime.now() > self.__expires:
-                self.__token, self.__session, self.__refresh_token, self.__expires = self._refresh()
-
-            return self.__token
-
-    @backoff.on_exception(backoff.constant, requests.exceptions.RequestException, max_time=300, giveup=fatal_code,
-                          interval=15)
-    def _refresh(self) -> Tuple[str, str, str, datetime]:
-        data = {
-            'client_id': 'cdse-public',
-            'refresh_token': self.__refresh_token,
-            'grant_type': 'refresh_token',
-        }
-
-        response = requests.post(DEFAULT_AUTH_ENDPOINT, data=data)
-
-        logger.info(f'Refresh Dataspace token request: {response.status_code}')
-
-        response.raise_for_status()
-
-        try:
-            response_json = response.json()
-
-            access_token = response_json['access_token']
-            session_id = response_json['session_state']
-            refresh_token = response_json['refresh_token']
-            expires_in = response_json['expires_in']
-        except KeyError as e:
-            raise RuntimeError(
-                f'Failed to parse expected field "{str(e)}" from authentication response.'
-            )
-
-        logger.info(f'Refreshed Dataspace token for session {self.__session}')
-
-        return access_token, session_id, refresh_token, datetime.now() + timedelta(seconds=expires_in)
-
-    @backoff.on_exception(backoff.constant, requests.exceptions.RequestException, max_time=300, giveup=fatal_code,
-                          interval=15)
-    def _get_token(self, username: str, password: str) -> Tuple[str, str, str, datetime]:
-        data = {
-            'client_id': 'cdse-public',
-            'username': username,
-            'password': password,
-            'grant_type': 'password'
-        }
-
-        response = requests.post(DEFAULT_AUTH_ENDPOINT, data=data)
-
-        logger.info(f'Get Dataspace token for {username}: {response.status_code}')
-
-        response.raise_for_status()
-
-        try:
-            response_json = response.json()
-
-            access_token = response_json['access_token']
-            session_id = response_json['session_state']
-            refresh_token = response_json['refresh_token']
-            expires_in = response_json['expires_in']
-        except KeyError as e:
-            raise RuntimeError(
-                f'Failed to parse expected field "{str(e)}" from authentication response.'
-            )
-
-        logger.info(f'Created Dataspace session {session_id}')
-
-        return access_token, session_id, refresh_token, datetime.now() + timedelta(seconds=expires_in)
-
-    @backoff.on_exception(backoff.constant, requests.exceptions.RequestException, max_time=300, giveup=fatal_code,
-                          interval=15)
-    def _delete_token(self):
-        url = f'{DEFAULT_SESSION_ENDPOINT}/{self.__session}'
-        headers = {'Authorization': f'Bearer {self.token}', 'Content-Type': 'application/json'}
-
-        response = requests.delete(url=url, headers=headers)
-
-        logger.info(f'Delete request {response.url}: {response.status_code}')
-        response.raise_for_status()
-
-    def __enter__(self):
-        return self
-
-    def __exit__(self, exc_type, exc_value, exc_traceback):
-        self._delete_token()
 
 
 def get_parser():
@@ -233,7 +113,12 @@ def build_query_filter(platform, *args):
     return {'$filter': filter_string, '$orderby': 'ContentDate/Start asc', '$top': str(QUERY_PAGE_SIZE)}
 
 
-@backoff.on_exception(backoff.constant, requests.exceptions.RequestException, max_time=300, giveup=fatal_code, interval=15)
+@backoff.on_exception(backoff.constant,
+                      requests.exceptions.RequestException,
+                      max_time=300,
+                      giveup=fatal_code,
+                      on_backoff=backoff_logger,
+                      interval=15)
 def _do_query(url, **kwargs):
     response = requests.get(url, **kwargs)
 
@@ -279,15 +164,19 @@ def main():
     query_results = query(build_query_filter(args.platform, *query_filters))
 
     if len(query_results) == 0:
-        raise NoQueryResultsException()
+        raise NoQueryResultsException('No results returned from parsed query results')
 
     if args.url_only:
         urls = {r['Name']: f'{DEFAULT_DOWNLOAD_ENDPOINT}({r["Id"]})/$value' for r in query_results}
         print(json.dumps(urls, indent=4))
     else:
         with DataspaceSession(args.username, args.password) as dss:
-            @backoff.on_exception(backoff.constant, requests.exceptions.RequestException, max_time=300,
-                                  giveup=fatal_code, interval=15)
+            @backoff.on_exception(backoff.constant,
+                                  requests.exceptions.RequestException,
+                                  max_time=300,
+                                  giveup=fatal_code,
+                                  on_backoff=backoff_logger,
+                                  interval=15)
             def do_download(gid, filename):
                 start_t = datetime.now()
